requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
boto3
python-multipart
pydantic-settings

